Some notification cleanup
